requirements.txt
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
psycopg2==2.9.6
